\name{florida_hurr20}
\alias{florida_hurr20}
\title{Hurricanes affecting Florida, USA from 1900-1999}
\description{
 Atlantic hurricanes recorded as affecting Florida, USA during the 20th century.
}
\usage{
 data(florida_hurr20)
}
\source{
 Blake, E.S., Rappaport, E.N. & Landsea, C.W. (2007)
 The deadliest, costliest, and most intense United States tropical
 cyclones from 1851 to 2006 (and other frequently requested hurricane facts).
 NOAA Technical Memorandum NWS TPC-5, Miami: National Weather Service.
}
\keyword{misc}
