\name{florida_temp20}
\alias{florida_temp20}
\title{Average temperatures in Florida, USA from 1900-1999}
\description{
 Average statewide temperatures in Florida, USA from 1900-1999.
}
\usage{
 data(florida_temp20)
}
\source{
 Florida Climate Center (2013)
 Florida statewide averaged temperature data (in degrees Fahrenheit).
 URL: http://climatecenter.fsu.edu/products-services/data/statewide-averages/temperature
 accessed 28/5/2013
}
\keyword{misc}
